requirements.txt
asgiref==3.7.2
Django==4.2.9
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.1
pillow==10.2.0
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.9.0
